+++ tests/testthat/test-cookieBanner.R
@@ -39,12 +39,12 @@ test_that("cookie banner link markup is as expected", {
 })
 
 
-test_that("width defaults to standard and doesn't add a wide/full class", {
+test_that("width defaults to standard, no three-quarters/full class", {
   cookie_banner_check <- cookieBanner("The best thing")
   container <- find_tag_required(cookie_banner_check, "govuk-width-container")
   expect_null(htmltools::tagGetAttribute(container, "style"))
   expect_no_tag(cookie_banner_check, "govuk-width-container--standard")
-  expect_no_tag(cookie_banner_check, "govuk-width-container--wide")
+  expect_no_tag(cookie_banner_check, "govuk-width-container--three-quarters")
   expect_no_tag(cookie_banner_check, "govuk-width-container--full")
 })